tolerance

+++ difficulty/difficulty.go
@@ -25,7 +25,7 @@ func CalcNextRequiredDifficulty(lastHeaderTime time.Time,difficulty *big.Int, ne
 	ParentDiff := difficulty
 
 	// Calc max(1 - (block_slot - parent_slot) // 10, -199)
-	MaxOne := new(big.Int).SetInt64(max(1-(newBlockTime.Unix()/poc.PoCSlot-lastHeaderTime.Unix()/int64(poc.PoCSlot))/int64(poc.ToleranceSlot), -199))
+	MaxOne := new(big.Int).SetInt64(max(1-(newBlockTime.Unix()/poc.PoCSlot-lastHeaderTime.Unix()/int64(poc.PoCSlot))/poc.ToleranceSlot, -199))
 
 	// Calc adjusted part, parent_diff // 2048 * MaxOne
 	Adjusted := new(big.Int).Mul(new(big.Int).Div(ParentDiff, diffUnit), MaxOne)
